requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pillow==10.4.0
sqlalchemy==2.0.36
alembic==1.13.2
python-multipart==0.0.9
pydantic-settings==2.4.0
httpx==0.27.2
pytest==8.3.2
loguru==0.7.2
# Optional for captioning:
transformers==4.44.2
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
